Fixed commission list actions.
Made it so the commission list actions only show up for the writing agents as far as edit, delete, chargeback etc.

Made it so chargebacks don't show up on items that are already chargebacks.

Made it so deleting a commission request will delete all the corresponding commissions from that writing agent.

Still have a bug on the chargeback where it's not getting the right agent information... not sure why.