Got chargeback and delete working on policies.
Implemented an overall actions processor for admin to be able to handle actions
outside of the affiliate-ltp mechanism.

Implemented a notices system that mimics the symfony flash concept with a one time use notice
that shows up using the wordpress transient system.

Policies chargeback and delete functionality setup and moved from the class-commissions.
Broke the commissions functaionlity up into a bunch of classes that's more in line with single reponsibility principle.